ModelLogic -> GameModel
