+++ demos/snake/snake/test/src/snake/model/GameModelTests.scala
@@ -1,4 +1,4 @@
-package snake.gamelogic
+package snake.model
 
 import indigo.{GameTime, Seconds}
 import indigo.shared.events.FrameTick
@@ -11,18 +11,18 @@ import indigo.shared.dice.Dice
 import indigoextras.geometry.BoundingBox
 import indigoextras.geometry.Vertex
 
-class ModelLogicTests extends munit.FunSuite {
+class GameModelTests extends munit.FunSuite {
 
   val defaultDelay: Seconds = Seconds(0.1)
 
   val model: GameModel =
-    ModelLogic.initialModel(
+    GameModel.initialModel(
       BoundingBox(0, 0, 5, 5), // grid square size 10
       ControlScheme.directedKeys
     )
 
   test("basic model updates should advance the game on frame tick") {
-    val actual = ModelLogic.update(GameTime.is(Seconds(0.15)), Dice.loaded(1), model, 10)(FrameTick).unsafeGet
+    val actual = GameModel.update(GameTime.is(Seconds(0.15)), Dice.loaded(1), model, 10)(FrameTick).unsafeGet
     val expected = model.copy(
       snake = model.snake.copy(start = Vertex(2, 2)),
       gameState = model.gameState.updateNow(Seconds(0.15), model.gameState.lastSnakeDirection)
